fix: fix output dir of audio
