+++ src/app.py
@@ -1,0 +1,8 @@
+﻿from gpt_sovits import get_audio 
+
+def main():
+    audio = get_audio("你好嗎？") # testing purposes
+    
+
+if __name__ == '__main__':
+    main()
